created left menu drawer, grant route
