+++ grantly/src/components/LeftDrawer.js
@@ -1,0 +1,65 @@
+import React from "react";
+import { adminStyles } from "../styles/adminStyles";
+import SwipeableDrawer from "@material-ui/core/SwipeableDrawer";
+import Button from "@material-ui/core/Button";
+import List from "@material-ui/core/List";
+import ListItem from "@material-ui/core/ListItem";
+import ListItemIcon from "@material-ui/core/ListItemIcon";
+import ListItemText from "@material-ui/core/ListItemText";
+import MenuIcon from "@material-ui/icons/Menu";
+import VerticalSplitIcon from "@material-ui/icons/VerticalSplit";
+import ListIcon from "@material-ui/icons/List";
+import Typography from "@material-ui/core/Typography";
+
+export default function LeftDrawer() {
+  const toggleDrawer = (side, open) => event => {
+    if (
+      event &&
+      event.type === "keydown" &&
+      (event.key === "Tab" || event.key === "Shift")
+    ) {
+      return;
+    }
+    setState({ ...state, [side]: open });
+  };
+
+  const sideList = side => (
+    <div
+      className={adminStyles.list}
+      role="presentation"
+      onClick={toggleDrawer(side, false)}
+      onKeyDown={toggleDrawer(side, false)}
+    >
+      <List>
+        {["Moderator Management","Detail View", "Table View"].map((text, index) => (
+          <ListItem button key={text}>
+            <ListItemIcon>
+              {index % 2 === 0 ? <VerticalSplitIcon /> : <ListIcon />}
+            </ListItemIcon>
+            <ListItemText primary={text} />
+          </ListItem>
+        ))}
+      </List>
+    </div>
+  );
+
+  const [state, setState] = React.useState({
+    left: false,
+
+  });
+
+  return (
+    <div>
+      <Button onClick={toggleDrawer("left", true)}>
+        <MenuIcon />
+      </Button>
+      <SwipeableDrawer
+        open={state.left}
+        onClose={toggleDrawer("left", false)}
+        onOpen={toggleDrawer("left", true)}
+      >
+        {sideList("left")}
+      </SwipeableDrawer>
+    </div>
+  );
+}
